DFPL-1323: merge urgent/gatekeeping eventData objects
