+++ service/src/main/java/uk/gov/hmcts/reform/fpl/controllers/AddUrgentDirectionsController.java
@@ -69,7 +69,7 @@ public class AddUrgentDirectionsController extends CallbackController {
 
         if (nonNull(caseData.getAllocatedJudge())) {
             data.put("gatekeepingOrderIssuingJudge", orderService.setAllocatedJudgeLabel(caseData.getAllocatedJudge(),
-                caseData.getUrgentDirectionsOrderEventData().getGatekeepingOrderIssuingJudge()));
+                caseData.getGatekeepingOrderEventData().getGatekeepingOrderIssuingJudge()));
         }
 
         return respond(data);
@@ -85,9 +85,9 @@ public class AddUrgentDirectionsController extends CallbackController {
                 "The direction-selection callback does not support %s route ", caseData.getUrgentDirectionsRouter()));
         }
 
-        orderService.populateUrgentDirections(caseDetails);
+        orderService.populateStandardDirections(caseDetails);
 
-        orderService.getHearing(caseDetails)
+        orderService.getHearing(caseData)
             .map(HearingBooking::getStartDate)
             .map(hearingDate -> formatLocalDateTimeBaseUsingFormat(hearingDate, DATE_TIME))
             .ifPresent(hearingDate -> {
@@ -102,13 +102,13 @@ public class AddUrgentDirectionsController extends CallbackController {
 
     @PostMapping("/prepare-decision/mid-event")
     public AboutToStartOrSubmitCallbackResponse handleGenerateDraftMidEvent(@RequestBody CallbackRequest request) {
-        final CaseData caseData = orderService.updateUrgentDirections(request.getCaseDetails());
+        final CaseData caseData = orderService.updateStandardDirections(request.getCaseDetails());
         final CaseDetailsMap data = caseDetailsMap(request.getCaseDetails());
 
         data.put("gatekeepingOrderSealDecision", orderService.buildSealDecision(caseData));
 
         if (caseData.getUrgentDirectionsRouter() == SERVICE) {
-            data.put("standardDirections", caseData.getUrgentDirectionsOrderEventData().getStandardDirections());
+            data.put("standardDirections", caseData.getGatekeepingOrderEventData().getStandardDirections());
         }
         return respond(data);
     }
@@ -116,17 +116,17 @@ public class AddUrgentDirectionsController extends CallbackController {
     @PostMapping("/about-to-submit")
     public AboutToStartOrSubmitCallbackResponse handleAboutToSubmit(@RequestBody CallbackRequest request) {
 
-        final CaseData caseData = orderService.updateUrgentDirections(request.getCaseDetails());
+        final CaseData caseData = orderService.updateStandardDirections(request.getCaseDetails());
         final CaseDetailsMap data = caseDetailsMap(request.getCaseDetails());
 
         final var sdoRouter = caseData.getUrgentDirectionsRouter();
 
         switch (sdoRouter) {
             case UPLOAD:
-                data.put("standardDirectionOrder", orderService.buildUrgentDirectionsOrderFromUploadedFile(caseData));
+                data.put("standardDirectionOrder", orderService.buildOrderFromUploadedFile(caseData));
                 break;
             case SERVICE:
-                data.put("standardDirectionOrder", orderService.buildUrgentDirectionsOrderFromGeneratedFile(caseData));
+                data.put("standardDirectionOrder", orderService.buildOrderFromGeneratedFile(caseData));
                 break;
         }
