+++ service/src/main/java/uk/gov/hmcts/reform/fpl/controllers/ListGatekeepingHearingController.java
@@ -112,7 +112,7 @@ public class ListGatekeepingHearingController extends CallbackController {
     public CallbackResponse handleAboutToSubmit(@RequestBody CallbackRequest callbackRequest) {
 
         final CaseDetails caseDetails = callbackRequest.getCaseDetails();
-        CaseData eventData = orderService.updateGatekeepingDirections(callbackRequest.getCaseDetails());
+        CaseData eventData = orderService.updateStandardDirections(callbackRequest.getCaseDetails());
         final CaseDetailsMap caseData = caseDetailsMap(caseDetails);
 
         hearingsService.findAndSetPreviousVenueId(eventData);
@@ -276,9 +276,9 @@ public class ListGatekeepingHearingController extends CallbackController {
         eventData = mergeEventAndCaseData(eventData, caseData);
 
         if (UPLOAD == sdoRouter) {
-            caseData.put("standardDirectionOrder", orderService.buildGatekeepingOrderFromUploadedFile(eventData));
+            caseData.put("standardDirectionOrder", orderService.buildOrderFromUploadedFile(eventData));
         } else if (SERVICE == sdoRouter) {
-            caseData.put("standardDirectionOrder", orderService.buildGatekeepingOrderFromGeneratedFile(eventData));
+            caseData.put("standardDirectionOrder", orderService.buildOrderFromGeneratedFile(eventData));
         }
 
         callbackRequest.getCaseDetails().setData(caseData);
