+++ service/src/main/java/uk/gov/hmcts/reform/fpl/model/event/GatekeepingOrderEventData.java
@@ -2,40 +2,112 @@ package uk.gov.hmcts.reform.fpl.model.event;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.Builder;
 import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.experimental.SuperBuilder;
 import lombok.extern.jackson.Jacksonized;
 import uk.gov.hmcts.reform.fpl.enums.DirectionType;
+import uk.gov.hmcts.reform.fpl.enums.LanguageTranslationRequirement;
+import uk.gov.hmcts.reform.fpl.enums.YesNo;
+import uk.gov.hmcts.reform.fpl.model.Allocation;
+import uk.gov.hmcts.reform.fpl.model.CustomDirection;
+import uk.gov.hmcts.reform.fpl.model.GatekeepingOrderSealDecision;
+import uk.gov.hmcts.reform.fpl.model.StandardDirection;
+import uk.gov.hmcts.reform.fpl.model.Temp;
+import uk.gov.hmcts.reform.fpl.model.common.DocumentReference;
+import uk.gov.hmcts.reform.fpl.model.common.Element;
+import uk.gov.hmcts.reform.fpl.model.common.JudgeAndLegalAdvisor;
 
+import java.lang.reflect.Field;
+import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.stream.Stream;
 
 import static java.util.stream.Collectors.toList;
+import static org.apache.commons.lang3.ObjectUtils.defaultIfNull;
+import static org.apache.commons.lang3.reflect.FieldUtils.getFieldsListWithAnnotation;
 
 @Data
-@SuperBuilder(toBuilder = true)
+@Builder(toBuilder = true)
 @Jacksonized
 @JsonInclude(JsonInclude.Include.NON_NULL)
-@EqualsAndHashCode(callSuper = true)
-public class GatekeepingOrderEventData extends OrderEventData {
+public class GatekeepingOrderEventData {
+
+    @Temp
+    DocumentReference urgentHearingOrderDocument;
+    @Temp
+    Allocation urgentHearingAllocation;
+    @Temp
+    YesNo showUrgentHearingAllocation;
+
+    JudgeAndLegalAdvisor gatekeepingOrderIssuingJudge;
+    GatekeepingOrderSealDecision gatekeepingOrderSealDecision;
+
     List<DirectionType> directionsForAllParties;
+    List<DirectionType> directionsForCourt;
+    List<DirectionType> directionsForCourtUpdated;
     List<DirectionType> directionsForLocalAuthority;
-    List<DirectionType> directionsForRespondents;
     List<DirectionType> directionsForCafcass;
     List<DirectionType> directionsForCafcassUpdated;
-    List<DirectionType> directionsForCourt;
+    List<DirectionType> directionsForRespondents;
     List<DirectionType> directionsForOthers;
+
+    List<DirectionType> urgentDirectionsForAllParties;
+    List<DirectionType> urgentDirectionsForLocalAuthority;
+    List<DirectionType> urgentDirectionsForCafcass;
+
+
+    List<Element<CustomDirection>> customDirections;
+    List<Element<StandardDirection>> standardDirections;
+
+    String gatekeepingOrderListOrSendToAdminReason;
+    String gatekeepingOrderListOrSendToAdmin;
+
+    LanguageTranslationRequirement gatekeepingTranslationRequirements;
+    LanguageTranslationRequirement urgentGatekeepingTranslationRequirements;
+
+    DocumentReference currentSDO;
+    YesNo useUploadRoute;
+    YesNo useServiceRoute;
 
     @JsonIgnore
     public List<DirectionType> getRequestedDirections() {
-        return Stream.of(directionsForAllParties, directionsForLocalAuthority,
-            directionsForRespondents, directionsForCafcass, directionsForCafcassUpdated,
-            directionsForOthers, directionsForCourt, directionsForCourtUpdated)
+        return Stream.of(urgentDirectionsForAllParties, urgentDirectionsForLocalAuthority, urgentDirectionsForCafcass,
+                directionsForAllParties, directionsForLocalAuthority, directionsForCafcass, directionsForCourt,
+                directionsForCafcassUpdated, directionsForCourtUpdated, directionsForRespondents, directionsForOthers)
             .filter(Objects::nonNull)
             .flatMap(Collection::stream)
             .collect(toList());
+    }
+
+    public JudgeAndLegalAdvisor getGatekeepingOrderIssuingJudge() {
+        return defaultIfNull(gatekeepingOrderIssuingJudge, JudgeAndLegalAdvisor.builder().build());
+    }
+
+    public GatekeepingOrderSealDecision getGatekeepingOrderSealDecision() {
+        return defaultIfNull(gatekeepingOrderSealDecision, GatekeepingOrderSealDecision.builder().build());
+    }
+
+    public static List<String> temporaryFields() {
+        return getFieldsListWithAnnotation(GatekeepingOrderEventData.class, Temp.class).stream()
+            .map(Field::getName)
+            .collect(toList());
+    }
+
+    public List<Element<StandardDirection>> resetStandardDirections() {
+        this.standardDirections = new ArrayList<>();
+        return standardDirections;
+    }
+
+    public boolean isSentToAdmin() {
+        return Optional.ofNullable(gatekeepingOrderListOrSendToAdmin)
+            .map(value -> value.equals("NO"))
+            .orElse(false);
+    }
+
+    public String getSendToAdminReason() {
+        return gatekeepingOrderListOrSendToAdminReason;
     }
 }
