+++ service/src/test/java/uk/gov/hmcts/reform/fpl/model/event/GatekeepingOrderEventDataTest.java
@@ -31,7 +31,7 @@ class GatekeepingOrderEventDataTest {
 
     @Test
     void shouldReturnRequestedStandardDirectionTypes() {
-        final OrderEventData underTest = GatekeepingOrderEventData.builder()
+        final GatekeepingOrderEventData underTest = GatekeepingOrderEventData.builder()
             .directionsForAllParties(List.of(REQUEST_PERMISSION_FOR_EXPERT_EVIDENCE, ASK_FOR_DISCLOSURE))
             .directionsForLocalAuthority(List.of(SEND_DOCUMENTS_TO_ALL_PARTIES, SEND_MISSING_ANNEX))
             .directionsForCafcass(List.of(APPOINT_CHILDREN_GUARDIAN))
@@ -69,7 +69,7 @@ class GatekeepingOrderEventDataTest {
 
     @Test
     void shouldResetStandardDirections() {
-        final OrderEventData underTest = GatekeepingOrderEventData.builder()
+        final GatekeepingOrderEventData underTest = GatekeepingOrderEventData.builder()
             .standardDirections(wrapElements(
                 StandardDirection.builder().build(),
                 StandardDirection.builder().build()))
@@ -81,7 +81,7 @@ class GatekeepingOrderEventDataTest {
 
     @Test
     void shouldReturnLanguageRequirements() {
-        final OrderEventData underTest = GatekeepingOrderEventData.builder()
+        final GatekeepingOrderEventData underTest = GatekeepingOrderEventData.builder()
             .gatekeepingTranslationRequirements(ENGLISH_TO_WELSH)
             .build();
